be able to refresh person (ie organization
