+++ tests/php-unit-tests/HybridAuthLoginExtensionTest.php
@@ -15,7 +15,6 @@ use URP_UserProfile;
 use UserExternal;
 use utils;
 
-
 /**
  *
  * @runTestsInSeparateProcesses
